requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
sentence-transformers==2.2.2
python-dotenv==1.0.0
psutil==5.9.6
pytest==7.4.3
faiss-cpu==1.7.4
numpy>=1.20.0,<1.24.0 